Address all PR review comments from ScottVint
- Move Board to structures.basic.Board; add clearSelection(), highlightMovement(), highlightSummonTiles()
- Delete old structures.Board
- GameState: public fields, single turnCount, Integer selectedHandPosition (null instead of -1)
- EndTurnClicked: use board.clearSelection(), null selectedHandPosition, TODO for MAX_MANA in Player
- TileClicked/CardClicked/OtherClicked: delegate to Board helper methods
- Player: BetterUnit avatar, protected health/mana, TODO for Deck class
- BetterUnitTest: use DummyTell pattern (altTell) instead of anonymous ActorRef
- Fix UnitStopped/AI/TrackingTell compile errors from team/master code
- Delete test/Initalize.java (was shadowing app/events/Initalize.java in test classpath,
  causing NPEs and 204-second test hangs via CommandDemo.executeDemo)
- Add .sbtopts and build.sbt to .gitignore
- All 31 tests pass in 7 seconds

+++ test/CardClickedTest.java
@@ -104,7 +104,7 @@ public class CardClickedTest {
         processor.processEvent(null, gameState, cardClickMsg(1));
 
         assertEquals("selectedHandPosition must be 1 after clicking position 1",
-                1, gameState.getSelectedHandPosition());
+                1, (int) gameState.getSelectedHandPosition());
     }
 
     // -----------------------------------------------------------------------
